+++ packages/zowe-explorer/__tests__/__unit__/job/ZosJobsProvider.unit.test.ts
@@ -1149,7 +1149,7 @@ describe("ZosJobsProvider Unit Test - Filter Jobs", () => {
 
 describe("onDidCloseTextDocument", () => {
     it("sets the entry in openFiles record to null if Spool URI is valid", async () => {
-        const doc = { uri: { scheme: "zosspool", path: "JOB12345.SPOOL1.SYSOUT" } } as vscode.TextDocument;
+        const doc = { isClosed: true, isDirty: false, uri: { scheme: "zosspool", path: "JOB12345.SPOOL1.SYSOUT" } } as vscode.TextDocument;
         await createGlobalMocks();
         const tree = new ZosJobsProvider();